change int to int64

+++ other/other.go
@@ -15,8 +15,8 @@ func GetDescription() string {
 	return sb.String()
 }
 
-func CheckLength(i int) {
-	if i > MaxBufferSize {
+func CheckLength(i int64) {
+	if i > int64(MaxBufferSize) {
 		panic(GetDescription())
 	}
 }
